pyproject.toml
[tool.poetry]
name = "multimodal-rag"
version = "0.1.0"
description = ""
authors = ["Christopher Keibel <christopher.keibel.90@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
transformers = "^4.40.1"
torch = "2.0.0"
pydantic = "^2.7.1"
gradio = "^4.29.0"
fastapi = "^0.111.0"
accelerate = "0.30.1"
loguru = "^0.7.2"
unstructured = {extras = ["all-docs"], version = "^0.15.0"}
pillow = "^10.2.0"
timm = "^1.0.8"
pymupdf = "^1.24.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
